move redux to predict components
