+++ apps/src/lab2/views/components/Instructions/PredictQuestion.tsx
@@ -3,10 +3,12 @@ import {RadioButton} from '@code-dot-org/component-library/radioButton';
 import classNames from 'classnames';
 import React from 'react';
 
+import {PredictQuestionType} from '@cdo/apps/lab2/levelEditors/types';
 import {
-  LevelPredictSettings,
-  PredictQuestionType,
-} from '@cdo/apps/lab2/levelEditors/types';
+  isPredictAnswerLocked,
+  setPredictResponse,
+} from '@cdo/apps/lab2/redux/predictLevelRedux';
+import {useAppDispatch, useAppSelector} from '@cdo/apps/util/reduxHooks';
 
 import {PREDICT_FREE_RESPONSE_DEFAULT_HEIGHT} from '../../../constants';
 
@@ -15,20 +17,19 @@ import PredictResetButton from './PredictResetButton';
 import moduleStyles from './predict.module.scss';
 
 interface PredictQuestionProps {
-  predictSettings: LevelPredictSettings | undefined;
-  predictResponse: string | undefined;
-  setPredictResponse: (response: string) => void;
-  predictAnswerLocked: boolean;
   className?: string;
 }
 
 const PredictQuestion: React.FunctionComponent<PredictQuestionProps> = ({
-  predictSettings,
-  predictResponse,
-  setPredictResponse,
-  predictAnswerLocked,
   className,
 }) => {
+  const predictSettings = useAppSelector(
+    state => state.lab.levelProperties?.predictSettings
+  );
+  const predictResponse = useAppSelector(state => state.predictLevel.response);
+  const predictAnswerLocked = useAppSelector(isPredictAnswerLocked);
+  const dispatch = useAppDispatch();
+
   if (!predictSettings?.isPredictLevel) {
     return null;
   }
@@ -42,9 +43,9 @@ const PredictQuestion: React.FunctionComponent<PredictQuestionProps> = ({
       } else if (selections.includes(value)) {
         selections.splice(selections.indexOf(value), 1);
       }
-      setPredictResponse(selections.join(','));
+      dispatch(setPredictResponse(selections.join(',')));
     } else {
-      setPredictResponse(value);
+      dispatch(setPredictResponse(value));
     }
   };
